+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to the "cph-ng" extension will be documented in this file.
 - **fix**: no ole when the output and answer are both empty
 - **feat**: add debounce for editor change event
 - **feat**: add stdoutThreshold configuration to control stdout output handling
+- **feat**: clear test case result
 
 ## 0.2.2